+++ Source/Acciaccatura.h
@@ -49,7 +49,7 @@ public:
    void PlayNote(NoteMessage note) override;
 
    void IntSliderUpdated(IntSlider* slider, int oldVal, double time) override {}
-   void FloatSliderUpdated(FloatSlider* slider, float oldVal, double time) override;
+   void FloatSliderUpdated(FloatSlider* slider, double oldVal, double time) override;
    void CheckboxUpdated(Checkbox* checkbox, double time) override;
    void DropdownUpdated(DropdownList* list, int oldVal, double time) override {}
 
@@ -80,9 +80,9 @@ private:
    IntSlider* mOffsetSlider{ nullptr };
    NoteMode mNoteMode{ NoteMode::Scale };
    DropdownList* mNoteModeDropdown{ nullptr };
-   float mHoldTimeMs{ 35.0f };
+   double mHoldTimeMs{ 35.0 };
    FloatSlider* mHoldTimeSlider{ nullptr };
-   float mGlideTimeMs{ 5.0f };
+   double mGlideTimeMs{ 5.0 };
    FloatSlider* mGlideTimeSlider{ nullptr };
 
    Modulations mModulation{ false };
